+++ src/asset_management/asset_loader.rs
@@ -4,7 +4,7 @@ use sdl2::image::LoadTexture;
 use sdl2::render::{Texture, TextureCreator};
 use sdl2::video::WindowContext;
 
-use super::{collider::ColliderAnimation, spriter_pro_collider_loader};
+use super::{cast_point::CastPoint, collider::ColliderAnimation, spriter_pro_collider_loader};
 
 
 pub fn load_texture<'a>(
@@ -49,11 +49,10 @@ fn look_for_textures<'a>(
     }
 }
 
-//TODO instead you could import the texture names from the scon file aswell
-pub fn load_anim_from_dir(dir: &str) -> Vec<(i32, String)> {
+pub fn load_anim_from_dir(dir: &str) -> Vec<(i64, String)> {
     let paths = fs::read_dir(dir).unwrap();
 
-    let mut vec: Vec<(i32, String)> = Vec::new();
+    let mut vec: Vec<(i64, String)> = Vec::new();
 
     let mut sprites_length = 0;
     for entry in paths {
@@ -68,11 +67,11 @@ pub fn load_anim_from_dir(dir: &str) -> Vec<(i32, String)> {
     vec
 }
 
-pub fn load_anim_and_data_from_dir(dir: &str) -> (Vec<(i32, String)>, Option<ColliderAnimation>) {
+pub fn load_anim_and_data_from_dir(dir: &str) -> (Vec<(i64, String)>, Option<ColliderAnimation>) {
     let paths = fs::read_dir(dir).unwrap();
 
-    let mut vec: Vec<(i32, String)> = Vec::new();
-    let mut data: Option<(Vec<i32>, ColliderAnimation)> = None;
+    let mut vec: Vec<(i64, String)> = Vec::new();
+    let mut data: Option<(Vec<i64>, ColliderAnimation)> = None;
 
     let mut sprites_length = 0;
     for entry in paths {
@@ -90,7 +89,7 @@ pub fn load_anim_and_data_from_dir(dir: &str) -> (Vec<(i32, String)>, Option<Col
     match data {
         Some(colliders) => {
             for i in 0..vec.len() {
-                vec[i].0 = (colliders.0[i] / 16) as i32;
+                vec[i].0 = colliders.0[i] / 16;
             }
             (vec, Some(colliders.1))
         }
